Review: Rename LoggingUtils.h
